computation / communication interleaving improvement
When there is only one node, then each worker directly outputs
the results (vs. using a point-to-point communication for sending
the results to the master process).

When there are more than one node, then a second master is created
in the same node where the global master resides, and results are
outputted the same exact way as before (in that node only). This way
the global master is released from the coordination of local sample
generation, hence it can coordinate the generation among other nodes.
Now the global master should not be the bottleneck.