Autonomous local masters (#2)
* initialization

* routine signature simplification and result accumulation

Common data needed by routines is now with global scope.

Resulting samples are accumulated first and later on outputted.
This is a precondition step we need before introducing super master topology.

* shm preparation and some pseudo-code

* right cmake conf for ubuntu

* global master / node master processing

* computation / communication interleaving improvement

When there is only one node, then each worker directly outputs
the results (vs. using a point-to-point communication for sending
the results to the master process).

When there are more than one node, then a second master is created
in the same node where the global master resides, and results are
outputted the same exact way as before (in that node only). This way
the global master is released from the coordination of local sample
generation, hence it can coordinate the generation among other nodes.
Now the global master should not be the bottleneck.

* fix processing with higher master-worker topology

* fix processing with higher master-worker topology

* improvement for the scenario with 1 node only

When there is only one node, then a secondary master is not needed,
meaning an extra MPI process is available for generating samples.

* Direct output and all-work mode in in master

Workers located in master node now directly send the
generated samples to the standard output.

Master also generates samples when there is no other
work to do. This happens in the master node only.

* reduce number of messages used for coordination

* deactivate sample generation at master level

* revert back mspar code and bring makefile

Makefile is preferred when installed CMake is lower than 3.5.1

* semi-autonomous scheduling

* fix shm rank gathering process

* missing bcast for number of nodes to all processes

* fix master-worker collection

* fix sending results to node master

* add diagnose option

* simplify code and remove dead code

* code simplification

* code documentation

* documentation and fix output